refactor: split utils file to constants
