+++ src/components/atoms/Text/utils.ts
@@ -1,18 +1,6 @@
 import S from './Text.module.scss';
+import { colorClassNames, sizeClassNames } from './constants';
 import { TextColors, TextSizes, TextStyles } from './types';
-
-const colorClassNames: Record<TextColors, string> = {
-  white: S.color_white,
-  grey: S.color_grey,
-  lightgrey: S.color_lightgrey,
-  black: S.color_black,
-};
-
-const sizeClassNames: Record<TextSizes, string> = {
-  default: S.text_default,
-  thin: S.text_thin,
-  lineheight: S.text_lineheight,
-};
 
 export const getColorClassName = (color: TextColors): string => {
   const className = colorClassNames[color];
